Fix: 비로그인 유저가 좋아요 버튼을 누를 경우, 로그인이 필요하다는 confirm 창이 뜨도록 수정

+++ src/components/common/PostCard.tsx
@@ -28,6 +28,7 @@ export default function PostCard({ post, isSearch = false }: PostCardProps) {
   );
   const getUser = useAllUserStore((state) => state.getUser);
   const loggedInUser = useAuthStore((state) => state.user);
+  const isLoggedIn = useAuthStore((state) => state.isLoggedIn);
 
   useEffect(() => {
     parsePostTitle();
@@ -106,11 +107,19 @@ export default function PostCard({ post, isSearch = false }: PostCardProps) {
     e: React.MouseEvent<HTMLButtonElement, MouseEvent>
   ) => {
     e.stopPropagation();
+    if (!isLoggedIn) {
+      const result = window.confirm(
+        "로그인이 필요합니다. 로그인 페이지로 이동하시겠습니까?"
+      );
+      if (result) navigate("/login");
+      else return;
+    }
+
     if (likeInformation) {
       const result = await deleteLike(likeInformation._id);
       if (result) {
         setLikeInformation(null);
-        setLikeCount((prev) => prev - 1);
+        setLikeCount((prev) => Math.max(prev - 1, 0));
       }
     } else {
       const result = await postLike(post._id);
